jwt-middleware/tests/requirements.txt
httpx==0.26.0
PyJWT==2.8.0
pytest==8.0.0
pytest-asyncio==0.23.5 